ipd_km_pipeline/requirements.txt
PyMuPDF>=1.23.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
pillow==10.4.0
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
scikit-image==0.24.0
shapely==2.0.4
pandas==2.2.2
matplotlib==3.9.1
seaborn==0.13.2
tqdm==4.66.4
pydantic==2.9.0
orjson==3.10.7
pyyaml==6.0.2
rich==13.7.1
pulp==2.8.0
